Refactor plugin into modular architecture for better maintainability
• Extract file explorer support to explorers.lua module
• Extract GitHub functionality to github.lua module
• Clean up main init.lua to focus on core functionality
• Update debug.lua to work with new modular structure
• Maintain all existing functionality with improved organization

Module structure:
- init.lua: Core copying logic, commands, keybindings
- explorers.lua: File explorer detection and path extraction
- github.lua: GitHub permalink generation and git operations
- debug.lua: Debug utilities and testing tools
